requirements.txt
requests
ttkthemes
Pillow
tkintermapview
python-dotenv
pandas
matplotlib
Flask-WTF
plotly
dash
dash-bootstrap-components
dash-mantine-components